Renamed test_utils.py to covidcast_test_base.py and moved it to src/common.
